--- a/docs/development/performance/index.rst
+++ b/docs/development/performance/index.rst
@@ -10,5 +10,6 @@ Performance is measured using different types of scenarios.
    :maxdepth: 1
    :caption: Further reading
 
+   profiling
    scenarios
    apachebench