--- a/requirements/dev.txt
+++ b/requirements/dev.txt
@@ -125,6 +125,7 @@ vng-api-common==1.0.41    # via -r requirements/ci.txt
 waitress==1.4.3           # via -r requirements/ci.txt, webtest
 webob==1.8.5              # via -r requirements/ci.txt, webtest
 webtest==2.0.33           # via -r requirements/ci.txt, django-webtest
+whitenoise==5.1.0         # via -r requirements/dev.in
 zgw-consumers==0.8.1      # via -r requirements/ci.txt
 
 # The following packages are considered to be unsafe in a requirements file: