--- a/src/openzaak/conf/dev.py
+++ b/src/openzaak/conf/dev.py
@@ -72,6 +72,13 @@ REST_FRAMEWORK["DEFAULT_RENDERER_CLASSES"] += (
     "rest_framework.renderers.BrowsableAPIRenderer",
 )
 
+#
+# DJANGO-SILK
+#
+if config("PROFILE", default=False):
+    INSTALLED_APPS += ["silk"]
+    MIDDLEWARE = ["silk.middleware.SilkyMiddleware"] + MIDDLEWARE
+
 warnings.filterwarnings(
     "error",
     r"DateTimeField .* received a naive datetime",